chore: merge eval_get* into eval.go/eval_test.go; AGENTS Go test guidance
- Fold GET decode/validate helpers into handler/eval.go
- Fold GET eval unit tests into eval_test.go; drop six micro-files
- AGENTS.md: t.Parallel() preference + docs/flagr_testing.md pointer
